user/serializers/profile_serializer.py
from rest_framework import serializers
from user.models.profile_model import *

class ProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = Profile
        fields = '__all__'

