requirements.txt
asgiref==3.2.10
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djoser==2.0.5
docutils==0.16
drf-yasg==1.17.1
geotext==0.4.0
googletrans==3.0.0
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==1.7.0
imutils==0.5.3
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
joblib==0.16.0
keyring==21.2.1
MarkupSafe==1.1.1
nltk==3.5
numpy==1.19.1
opencv-contrib-python==4.3.0.36
packaging==20.4
pandas==1.0.5
Pillow==7.2.0
pkginfo==1.5.0.1
psycopg2==2.8.6
Pygments==2.6.1
pyparsing==2.4.7
pytesseract==0.3.4
python-dateutil==2.8.1
pytz==2020.1
pywin32-ctypes==0.2.0
readme-renderer==26.0
regex==2020.7.14
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sniffio==1.1.0
sqlparse==0.3.1
textblob==0.15.3
tqdm==4.48.0
twine==3.2.0
uritemplate==3.0.1
urllib3==1.25.10
webencodings==0.5.1
zipp==3.1.0
